refactor: reply entries into their own component
