+++ resources/js/Pages/Messages/Show.vue
@@ -2,13 +2,9 @@
 import AppLayout from "../../Layouts/AppLayout.vue";
 import Card from "../../Components/Card.vue";
 import { router, useForm, usePage } from "@inertiajs/vue3";
-import { computed, ref } from "vue";
+import { computed } from "vue";
 import Reload from "../../Components/Icons/Reload.vue";
-import Trash from "../../Components/Icons/Trash.vue";
-import Pencil from "../../Components/Icons/Pencil.vue";
-import SecondaryButton from "../../Components/SecondaryButton.vue";
-import DangerButton from "../../Components/DangerButton.vue";
-import DialogModal from "../../Components/DialogModal.vue";
+import ReplyEntry from "../../Components/Messages/ReplyEntry.vue";
 import MessageData = App.Data.MessageData;
 import ReplyData = App.Data.ReplyData;
 
@@ -19,8 +15,6 @@ interface Props {
 const props = defineProps<Props>();
 
 const user = computed(() => usePage().props.auth.user);
-const confirmingReplyDeletion = ref(false);
-const selectedReply = ref(null);
 
 const form = useForm({
     content: "",
@@ -37,26 +31,6 @@ function submit() {
         },
     });
 }
-
-function deleteReply() {
-    router.delete(route("message.reply.destroy", selectedReply.value), {
-        preserveScroll: true,
-        onFinish: () => {
-            closeModal();
-            selectedReply.value = null;
-        },
-    });
-}
-
-function confirmReplyDeletion(reply: ReplyData) {
-    confirmingReplyDeletion.value = true;
-    selectedReply.value = reply;
-}
-
-const closeModal = () => {
-    confirmingReplyDeletion.value = false;
-    selectedReply.value = null;
-};
 </script>
 
 <template>
@@ -123,61 +97,12 @@ const closeModal = () => {
                         <template
                             v-for="reply in message.replies"
                             :key="reply.id">
-                            <div class="rounded-md bg-gray-700 p-2">
-                                <p>
-                                    {{ reply.content }}
-                                </p>
-
-                                <div class="flex justify-between gap-6">
-                                    <span
-                                        v-if="reply.user"
-                                        class="text-xs text-white/50">
-                                        Sent by: {{ reply.user.name }} - {{ reply.createdAt }}
-                                    </span>
-
-                                    <div
-                                        v-if="reply.userId === user.id"
-                                        class="flex gap-2">
-                                        <button class="transition-all hover:scale-105">
-                                            <Pencil class="size-4" />
-                                        </button>
-
-                                        <button
-                                            class="transition-all hover:scale-105 hover:text-red-400"
-                                            @click="confirmReplyDeletion(reply)">
-                                            <Trash class="size-4" />
-                                        </button>
-                                    </div>
-                                </div>
-                            </div>
+                            <ReplyEntry :reply="reply" />
                         </template>
                     </div>
                 </Card>
             </div>
         </div>
-
-        <DialogModal
-            :show="confirmingReplyDeletion"
-            @close="closeModal">
-            <template #title> Delete Reply</template>
-
-            <template #content>
-                Are you sure you want to delete your reply? Once deleted, we won't be able to restore any data related
-                to it. Please enter your password to confirm you would like to permanently delete your reply.
-            </template>
-
-            <template #footer>
-                <SecondaryButton @click="closeModal"> Cancel</SecondaryButton>
-
-                <DangerButton
-                    class="ms-3"
-                    :class="{ 'opacity-25': form.processing }"
-                    :disabled="form.processing"
-                    @click="deleteReply">
-                    Delete reply
-                </DangerButton>
-            </template>
-        </DialogModal>
     </AppLayout>
 </template>
